WIP: Added recipient_id to tweets
There is an error with SQLite that makes it so you cannot add a new
column to a table and specify null: false on it. It would normally make sense to
raise an error if we were adding null:false to a new column because
doing that in a database with entries would create rows which default to
null, but then violate the null: false. SQLite raises this exception
though regardless if there are records in the database or not. The work
around is to create a new column and then immediately change that
column. For more info see: http://stackoverflow.com/questions/3170634/how-to-solve-cannot-add-a-not-null-column-with-default-value-null-in-sqlite3